requirements.txt
transformers
cupy-cuda12x
numba
tqdm
numpy<2.0